modularization of functions
