+++ augmentation_ablation.py
@@ -3,22 +3,14 @@
 """
 
 import argparse
+from config import INPUT_SHAPE, NUM_SPLITS, RANDOM_SEED
 from dataset import generate_dataset, augmentations_order
 from sklearn.model_selection import StratifiedKFold
 import numpy as np
 import wandb
 from wandb.keras import WandbCallback
-from tensorflow_addons.optimizers import TriangularCyclicalLearningRate
-from tensorflow_addons.optimizers import SGDW
-from tensorflow.keras.optimizers import SGD
-from tensorflow.keras.metrics import TopKCategoricalAccuracy
-from tensorflow.keras.layers import Dropout, Dense
-from tensorflow.keras import Input
-from tensorflow.keras.models import Model
-from tensorflow.keras.applications.densenet import DenseNet121
-from tensorflow.keras.applications.densenet import preprocess_input as densenet121_preprocess_input
-from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2
-from tensorflow.keras.applications.mobilenet_v2 import preprocess_input as mobilenetv2_preprocess_input
+from model import build_densenet121_model, build_mobilenetv2_model
+from optimizer import build_sgd_optimizer
 import tensorflow as tf
 import pandas as pd
 
@@ -27,10 +19,6 @@ import pandas as pd
 # !wget "https://storage.googleapis.com/cloud-ai-platform-f3305919-42dc-47f1-82cf-4f1a3202db74/wlasl100_skeletons_train.csv" -nc
 # !wget "https://storage.googleapis.com/cloud-ai-platform-f3305919-42dc-47f1-82cf-4f1a3202db74/wlasl100_skeletons_val.csv" -nc
 # !wget "https://storage.googleapis.com/cloud-ai-platform-f3305919-42dc-47f1-82cf-4f1a3202db74/wlasl100_skeletons_test.csv" -nc
-
-NUM_SPLITS = 5
-INPUT_SHAPE = [None, 181, 3]
-RANDOM_SEED = 0
 
 # Load data
 train_dataframe = pd.read_csv("wlasl100_skeletons_train.csv", index_col=0)
@@ -47,83 +35,6 @@ labels = train_and_validation_dataframe.groupby(
     "video")["label"].unique().tolist()
 splits = list(skf.split(np.zeros(num_total_examples), labels))
 num_train_examples = len(splits[0][0])
-
-
-def build_sgd_optimizer(initial_learning_rate=0.001,
-                        maximal_learning_rate=0.01,
-                        step_size=50, momentum=0.0,
-                        nesterov=False, weight_decay=1e-7):
-    # setup schedule
-    learning_rate_schedule = TriangularCyclicalLearningRate(
-        initial_learning_rate=initial_learning_rate,
-        maximal_learning_rate=maximal_learning_rate,
-        step_size=step_size)
-
-    # setup the optimizer
-    if weight_decay:
-        initial_weight_decay = weight_decay
-        maximal_weight_decay = weight_decay * \
-            (maximal_learning_rate / initial_learning_rate)
-        weight_decay_schedule = TriangularCyclicalLearningRate(
-            initial_learning_rate=initial_weight_decay,
-            maximal_learning_rate=maximal_weight_decay,
-            step_size=step_size)
-
-        optimizer = SGDW(learning_rate=learning_rate_schedule,
-                         weight_decay=weight_decay_schedule,
-                         momentum=momentum, nesterov=nesterov)
-    else:
-        optimizer = SGD(learning_rate=learning_rate_schedule,
-                        momentum=momentum, nesterov=nesterov)
-    return optimizer
-
-
-def build_densenet121_model(input_shape=[None, 181, 3], dropout=0,
-                            optimizer=None, pretraining=True):
-    # setup model
-    weights = 'imagenet' if pretraining else None
-    inputs = Input(shape=input_shape)
-    inputs = densenet121_preprocess_input(inputs)
-    x = DenseNet121(input_shape=input_shape, weights=weights,
-                    include_top=False, pooling='avg')(inputs)
-    x = Dropout(dropout)(x)
-    predictions = Dense(100, activation='softmax')(x)
-    model = Model(inputs=inputs, outputs=predictions)
-
-    # setup the metrics
-    metrics = [
-        TopKCategoricalAccuracy(k=1, name='top_1', dtype=tf.float32)
-    ]
-
-    # compile the model
-    model.compile(optimizer=optimizer, loss='categorical_crossentropy',
-                  metrics=metrics)
-
-    return model
-
-
-def build_mobilenetv2_model(input_shape=[None, 181, 3], dropout=0,
-                            optimizer=None, pretraining=True):
-    # setup model
-    weights = "imagenet" if pretraining else None
-    inputs = Input(shape=input_shape)
-    inputs = mobilenetv2_preprocess_input(inputs)
-    x = MobileNetV2(input_shape=input_shape, weights=weights,
-                    include_top=False, pooling="avg")(inputs)
-    x = Dropout(dropout)(x)
-    predictions = Dense(100, activation='softmax')(x)
-    model = Model(inputs=inputs, outputs=predictions)
-
-    # setup the metrics
-    metrics = [
-        TopKCategoricalAccuracy(k=1, name='top_1', dtype=tf.float32)
-    ]
-
-    # compile the model
-    model.compile(optimizer=optimizer, loss='categorical_crossentropy',
-                  metrics=metrics)
-
-    return model
 
 
 def run_experiment(config=None, log_to_wandb=True, verbose=0):
