+++ config.py
@@ -1,0 +1,3 @@
+INPUT_SHAPE = [None, 181, 3]
+NUM_SPLITS = 5
+RANDOM_SEED = 0
